reorganise storage and catalog_manager into data_manager module
