+++ Cargo.toml
@@ -5,6 +5,6 @@ members = [
     "src/protocol",
     "src/sql_engine",
     "src/sql_types",
-    "src/storage",
+    "src/data_manager",
     "src/representation",
 ]
