+++ src/data_manager/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "storage"
+name = "data_manager"
 version = "0.1.0"
 authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
 edition = "2018"
