+++ src/data_manager/src/persistent.rs
