+++ src/data_manager/src/tests/manager/mod.rs
@@ -15,6 +15,25 @@
 use super::*;
 
 #[cfg(test)]
-mod delete;
+mod persistence;
 #[cfg(test)]
-mod select;
+mod queries;
+#[cfg(test)]
+mod schema;
+#[cfg(test)]
+mod table;
+
+const SCHEMA: &str = "schema_name";
+const SCHEMA_1: &str = "schema_name_1";
+const SCHEMA_2: &str = "schema_name_2";
+
+#[rstest::fixture]
+fn data_manager() -> DataManager {
+    DataManager::default()
+}
+
+#[rstest::fixture]
+fn data_manager_with_schema(data_manager: DataManager) -> DataManager {
+    data_manager.create_schema(SCHEMA).expect("schema is created");
+    data_manager
+}
