+++ src/data_manager/src/tests/manager/queries.rs
@@ -13,15 +13,13 @@
 // limitations under the License.
 
 use super::*;
-use representation::Datum;
+use representation::{Binary, Datum};
 use sql_types::SqlType;
 
 #[rstest::rstest]
-fn delete_all_from_table(catalog_manager_with_schema: CatalogManager) {
-    let schema_id = catalog_manager_with_schema
-        .schema_exists(SCHEMA)
-        .expect("schema exists");
-    let table_id = catalog_manager_with_schema
+fn delete_all_from_table(data_manager_with_schema: DataManager) {
+    let schema_id = data_manager_with_schema.schema_exists(SCHEMA).expect("schema exists");
+    let table_id = data_manager_with_schema
         .create_table(
             schema_id,
             "table_name",
@@ -32,7 +30,7 @@ fn delete_all_from_table(catalog_manager_with_schema: CatalogManager) {
         )
         .expect("table is created");
 
-    catalog_manager_with_schema
+    data_manager_with_schema
         .write_into(
             schema_id,
             table_id,
@@ -42,7 +40,7 @@ fn delete_all_from_table(catalog_manager_with_schema: CatalogManager) {
             )],
         )
         .expect("values are inserted");
-    catalog_manager_with_schema
+    data_manager_with_schema
         .write_into(
             schema_id,
             table_id,
@@ -52,7 +50,7 @@ fn delete_all_from_table(catalog_manager_with_schema: CatalogManager) {
             )],
         )
         .expect("values are inserted");
-    catalog_manager_with_schema
+    data_manager_with_schema
         .write_into(
             table_id,
             schema_id,
@@ -64,7 +62,7 @@ fn delete_all_from_table(catalog_manager_with_schema: CatalogManager) {
         .expect("values are inserted");
 
     assert_eq!(
-        catalog_manager_with_schema.delete_from(
+        data_manager_with_schema.delete_from(
             schema_id,
             table_id,
             vec![
@@ -77,9 +75,58 @@ fn delete_all_from_table(catalog_manager_with_schema: CatalogManager) {
     );
 
     assert_eq!(
-        catalog_manager_with_schema
+        data_manager_with_schema
             .full_scan(schema_id, table_id)
             .map(|iter| iter.map(Result::unwrap).map(Result::unwrap).collect()),
         Ok(vec![])
+    );
+}
+
+#[rstest::fixture]
+fn with_small_ints_table(data_manager_with_schema: DataManager) -> DataManager {
+    let schema_id = data_manager_with_schema.schema_exists(SCHEMA).expect("schema exists");
+    data_manager_with_schema
+        .create_table(
+            schema_id,
+            "table_name",
+            &[
+                ColumnDefinition::new("column_1", SqlType::SmallInt(i16::min_value())),
+                ColumnDefinition::new("column_2", SqlType::SmallInt(i16::min_value())),
+                ColumnDefinition::new("column_3", SqlType::SmallInt(i16::min_value())),
+            ],
+        )
+        .expect("table is created");
+    data_manager_with_schema
+}
+
+#[rstest::rstest]
+fn select_all_from_table_with_many_columns(with_small_ints_table: DataManager) {
+    let full_table_id = with_small_ints_table
+        .table_exists(SCHEMA, "table_name")
+        .expect("schema exists");
+    let schema_id = full_table_id.0;
+    let table_id = full_table_id.1.expect("table exist");
+    with_small_ints_table
+        .write_into(
+            schema_id,
+            table_id,
+            vec![(
+                Binary::pack(&[Datum::from_u64(1)]),
+                Binary::pack(&[Datum::from_i16(1), Datum::from_i16(2), Datum::from_i16(3)]),
+            )],
+        )
+        .expect("values are inserted");
+
+    assert_eq!(
+        with_small_ints_table.full_scan(schema_id, table_id).map(|read| read
+            .map(Result::unwrap)
+            .map(Result::unwrap)
+            .map(|(_key, values)| values)
+            .collect()),
+        Ok(vec![Binary::pack(&[
+            Datum::from_i16(1),
+            Datum::from_i16(2),
+            Datum::from_i16(3)
+        ])])
     );
 }
