+++ src/data_manager/tests/failpoints/drop_db.rs
@@ -12,11 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use data_manager::{Database, StorageError};
 use fail::FailScenario;
-use storage::{Database, DefinitionError, PersistentDatabase, StorageError};
 
 mod common;
 use common::{scenario, OBJECT, SCHEMA};
+use data_manager::persistent::PersistentDatabase;
 
 #[rstest::fixture]
 fn database() -> PersistentDatabase {
@@ -30,26 +31,29 @@ fn database() -> PersistentDatabase {
     storage
         .create_object(SCHEMA, OBJECT)
         .expect("no io error")
-        .expect("no storage error")
+        .expect("no platform errors")
         .expect("to create object");
     storage
 }
 
 #[rstest::rstest]
 fn io_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-tree", "return(io)").unwrap();
+    fail::cfg("sled-fail-to-drop-db", "return(io)").unwrap();
 
-    assert!(matches!(database.drop_object(SCHEMA, OBJECT), Err(_)));
+    assert_eq!(
+        database.drop_schema(SCHEMA).expect("no io error"),
+        Err(StorageError::CascadeIo(vec![OBJECT.to_owned()]))
+    );
 
     scenario.teardown();
 }
 
 #[rstest::rstest]
 fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-tree", "return(corruption)").unwrap();
+    fail::cfg("sled-fail-to-drop-db", "return(corruption)").unwrap();
 
     assert_eq!(
-        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
+        database.drop_schema(SCHEMA).expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -58,10 +62,10 @@ fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-tree", "return(bug)").unwrap();
+    fail::cfg("sled-fail-to-drop-db", "return(bug)").unwrap();
 
     assert_eq!(
-        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
+        database.drop_schema(SCHEMA).expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -69,11 +73,20 @@ fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
 }
 
 #[rstest::rstest]
+fn unsupported_operation_core_structure(database: PersistentDatabase, scenario: FailScenario) {
+    fail::cfg("sled-fail-to-drop-db", "return(unsupported(core_structure))").unwrap();
+
+    assert_eq!(database.drop_schema(SCHEMA).expect("no io error"), Ok(Ok(())));
+
+    scenario.teardown();
+}
+
+#[rstest::rstest]
 fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-tree", "return(unsupported)").unwrap();
+    fail::cfg("sled-fail-to-drop-db", "return(unsupported)").unwrap();
 
     assert_eq!(
-        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
+        database.drop_schema(SCHEMA).expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -82,11 +95,11 @@ fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn collection_not_found(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-tree", "return(collection_not_found)").unwrap();
+    fail::cfg("sled-fail-to-drop-db", "return(collection_not_found)").unwrap();
 
     assert_eq!(
-        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
-        Ok(Err(DefinitionError::ObjectDoesNotExist))
+        database.drop_schema(SCHEMA).expect("no io error"),
+        Err(StorageError::Storage)
     );
 
     scenario.teardown();
