+++ src/data_manager/tests/failpoints/drop_tree.rs
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use data_manager::{Database, DefinitionError, StorageError};
 use fail::FailScenario;
-use representation::Binary;
-use storage::{Database, DefinitionError, PersistentDatabase, StorageError};
 
 mod common;
 use common::{scenario, OBJECT, SCHEMA};
+use data_manager::persistent::PersistentDatabase;
 
 #[rstest::fixture]
 fn database() -> PersistentDatabase {
@@ -31,39 +31,26 @@ fn database() -> PersistentDatabase {
     storage
         .create_object(SCHEMA, OBJECT)
         .expect("no io error")
-        .expect("no platform errors")
+        .expect("no storage error")
         .expect("to create object");
     storage
 }
 
 #[rstest::rstest]
 fn io_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-insert-into-tree", "return(io)").unwrap();
+    fail::cfg("sled-fail-to-drop-tree", "return(io)").unwrap();
 
-    assert!(matches!(
-        database.write(
-            SCHEMA,
-            OBJECT,
-            vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
-        ),
-        Err(_)
-    ));
+    assert!(matches!(database.drop_object(SCHEMA, OBJECT), Err(_)));
 
     scenario.teardown();
 }
 
 #[rstest::rstest]
 fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-insert-into-tree", "return(corruption)").unwrap();
+    fail::cfg("sled-fail-to-drop-tree", "return(corruption)").unwrap();
 
     assert_eq!(
-        database
-            .write(
-                SCHEMA,
-                OBJECT,
-                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
-            )
-            .expect("no io error"),
+        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -72,16 +59,10 @@ fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-insert-into-tree", "return(bug)").unwrap();
+    fail::cfg("sled-fail-to-drop-tree", "return(bug)").unwrap();
 
     assert_eq!(
-        database
-            .write(
-                SCHEMA,
-                OBJECT,
-                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
-            )
-            .expect("no io error"),
+        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -90,16 +71,10 @@ fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-insert-into-tree", "return(unsupported)").unwrap();
+    fail::cfg("sled-fail-to-drop-tree", "return(unsupported)").unwrap();
 
     assert_eq!(
-        database
-            .write(
-                SCHEMA,
-                OBJECT,
-                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
-            )
-            .expect("no io error"),
+        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -108,16 +83,10 @@ fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn collection_not_found(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-insert-into-tree", "return(collection_not_found)").unwrap();
+    fail::cfg("sled-fail-to-drop-tree", "return(collection_not_found)").unwrap();
 
     assert_eq!(
-        database
-            .write(
-                SCHEMA,
-                OBJECT,
-                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
-            )
-            .expect("no io error"),
+        database.drop_object(SCHEMA, OBJECT).expect("no io error"),
         Ok(Err(DefinitionError::ObjectDoesNotExist))
     );
