+++ src/data_manager/tests/failpoints/flush_tree.rs
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use data_manager::{Database, DefinitionError, StorageError};
 use fail::FailScenario;
 use representation::Binary;
-use storage::{Database, DefinitionError, PersistentDatabase, StorageError};
 
 mod common;
-use crate::common::OBJECT;
-use common::{scenario, SCHEMA};
+use common::{scenario, OBJECT, SCHEMA};
+use data_manager::persistent::PersistentDatabase;
 
 #[rstest::fixture]
 fn database() -> PersistentDatabase {
@@ -30,7 +30,7 @@ fn database() -> PersistentDatabase {
         .expect("no platform errors")
         .expect("to create schema");
     storage
-        .create_object("schema_name", OBJECT)
+        .create_object(SCHEMA, OBJECT)
         .expect("no io error")
         .expect("no platform errors")
         .expect("to create object");
@@ -39,10 +39,14 @@ fn database() -> PersistentDatabase {
 
 #[rstest::rstest]
 fn io_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-remove-from-tree", "return(io)").unwrap();
+    fail::cfg("sled-fail-to-flush-tree", "return(io)").unwrap();
 
     assert!(matches!(
-        database.delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])]),
+        database.write(
+            SCHEMA,
+            OBJECT,
+            vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
+        ),
         Err(_)
     ));
 
@@ -51,11 +55,15 @@ fn io_error(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-remove-from-tree", "return(corruption)").unwrap();
+    fail::cfg("sled-fail-to-flush-tree", "return(corruption)").unwrap();
 
     assert_eq!(
         database
-            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .write(
+                SCHEMA,
+                OBJECT,
+                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
+            )
             .expect("no io error"),
         Err(StorageError::Storage)
     );
@@ -65,11 +73,15 @@ fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-remove-from-tree", "return(bug)").unwrap();
+    fail::cfg("sled-fail-to-flush-tree", "return(bug)").unwrap();
 
     assert_eq!(
         database
-            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .write(
+                SCHEMA,
+                OBJECT,
+                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
+            )
             .expect("no io error"),
         Err(StorageError::Storage)
     );
@@ -79,11 +91,15 @@ fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-remove-from-tree", "return(unsupported)").unwrap();
+    fail::cfg("sled-fail-to-flush-tree", "return(unsupported)").unwrap();
 
     assert_eq!(
         database
-            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .write(
+                SCHEMA,
+                OBJECT,
+                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
+            )
             .expect("no io error"),
         Err(StorageError::Storage)
     );
@@ -93,11 +109,15 @@ fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn collection_not_found(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-remove-from-tree", "return(collection_not_found)").unwrap();
+    fail::cfg("sled-fail-to-flush-tree", "return(collection_not_found)").unwrap();
 
     assert_eq!(
         database
-            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .write(
+                SCHEMA,
+                OBJECT,
+                vec![(Binary::with_data(vec![]), Binary::with_data(vec![]))]
+            )
             .expect("no io error"),
         Ok(Err(DefinitionError::ObjectDoesNotExist))
     );
