+++ src/data_manager/tests/failpoints/remove_from_tree.rs
@@ -12,11 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use data_manager::{Database, DefinitionError, StorageError};
 use fail::FailScenario;
-use storage::{Database, PersistentDatabase, StorageError};
+use representation::Binary;
 
 mod common;
-use common::{scenario, OBJECT, SCHEMA};
+use crate::common::OBJECT;
+use common::{scenario, SCHEMA};
+use data_manager::persistent::PersistentDatabase;
 
 #[rstest::fixture]
 fn database() -> PersistentDatabase {
@@ -28,7 +31,7 @@ fn database() -> PersistentDatabase {
         .expect("no platform errors")
         .expect("to create schema");
     storage
-        .create_object(SCHEMA, OBJECT)
+        .create_object("schema_name", OBJECT)
         .expect("no io error")
         .expect("no platform errors")
         .expect("to create object");
@@ -37,22 +40,24 @@ fn database() -> PersistentDatabase {
 
 #[rstest::rstest]
 fn io_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-db", "return(io)").unwrap();
+    fail::cfg("sled-fail-to-remove-from-tree", "return(io)").unwrap();
 
-    assert_eq!(
-        database.drop_schema(SCHEMA).expect("no io error"),
-        Err(StorageError::CascadeIo(vec![OBJECT.to_owned()]))
-    );
+    assert!(matches!(
+        database.delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])]),
+        Err(_)
+    ));
 
     scenario.teardown();
 }
 
 #[rstest::rstest]
 fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-db", "return(corruption)").unwrap();
+    fail::cfg("sled-fail-to-remove-from-tree", "return(corruption)").unwrap();
 
     assert_eq!(
-        database.drop_schema(SCHEMA).expect("no io error"),
+        database
+            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -61,10 +66,12 @@ fn corruption_error(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-db", "return(bug)").unwrap();
+    fail::cfg("sled-fail-to-remove-from-tree", "return(bug)").unwrap();
 
     assert_eq!(
-        database.drop_schema(SCHEMA).expect("no io error"),
+        database
+            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -72,20 +79,13 @@ fn reportable_bug(database: PersistentDatabase, scenario: FailScenario) {
 }
 
 #[rstest::rstest]
-fn unsupported_operation_core_structure(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-db", "return(unsupported(core_structure))").unwrap();
-
-    assert_eq!(database.drop_schema(SCHEMA).expect("no io error"), Ok(Ok(())));
-
-    scenario.teardown();
-}
-
-#[rstest::rstest]
 fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-db", "return(unsupported)").unwrap();
+    fail::cfg("sled-fail-to-remove-from-tree", "return(unsupported)").unwrap();
 
     assert_eq!(
-        database.drop_schema(SCHEMA).expect("no io error"),
+        database
+            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .expect("no io error"),
         Err(StorageError::Storage)
     );
 
@@ -94,11 +94,13 @@ fn unsupported_operation(database: PersistentDatabase, scenario: FailScenario) {
 
 #[rstest::rstest]
 fn collection_not_found(database: PersistentDatabase, scenario: FailScenario) {
-    fail::cfg("sled-fail-to-drop-db", "return(collection_not_found)").unwrap();
+    fail::cfg("sled-fail-to-remove-from-tree", "return(collection_not_found)").unwrap();
 
     assert_eq!(
-        database.drop_schema(SCHEMA).expect("no io error"),
-        Err(StorageError::Storage)
+        database
+            .delete(SCHEMA, OBJECT, vec![Binary::with_data(vec![])])
+            .expect("no io error"),
+        Ok(Err(DefinitionError::ObjectDoesNotExist))
     );
 
     scenario.teardown();
