+++ src/node/Cargo.toml
@@ -16,7 +16,7 @@ path = "src/bin.rs"
 kernel = { path = "../kernel" }
 sql_engine = { path = "../sql_engine" }
 sql_types = { path = "../sql_types" }
-storage = { path = "../storage" }
+data_manager = { path = "../data_manager" }
 futures-lite = "1.0.0"
 smol = "0.4.3"
 async-io = "0.2.7"
