+++ src/node/src/lib.rs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+extern crate data_manager;
 extern crate kernel;
 extern crate log;
 extern crate protocol;
-extern crate storage;
 
 pub mod node;
