+++ src/sql_engine/Cargo.toml
@@ -8,14 +8,13 @@ publish = false
 [dependencies]
 log = "0.4.11"
 kernel = { path = "../kernel" }
-storage = { path = "../storage" }
+data_manager = { path = "../data_manager" }
 sqlparser = { version = "0.6.1", features = ["bigdecimal"] }
 bigdecimal = "0.1.2"
 sql_types = { path = "../sql_types" }
 protocol = { path = "../protocol" }
 futures-lite = "1.0.0"
 representation = { path = "../representation" }
-serde = { version = "1.0.115", features = ["derive"] }
 bincode = "1.3.1"
 itertools = "0.9.0"
