+++ src/sql_engine/src/lib.rs
@@ -16,7 +16,6 @@ extern crate bigdecimal;
 extern crate log;
 
 use crate::{
-    catalog_manager::CatalogManager,
     ddl::{
         create_schema::CreateSchemaCommand, create_table::CreateTableCommand, drop_schema::DropSchemaCommand,
         drop_table::DropTableCommand,
@@ -25,6 +24,7 @@ use crate::{
     query::{bind::ParamBinder, plan::Plan, process::QueryProcessor},
     session::{statement::PreparedStatement, Session},
 };
+use data_manager::DataManager;
 use itertools::izip;
 use kernel::SystemResult;
 use protocol::{
@@ -34,8 +34,6 @@ use protocol::{
     sql_values::PostgreSqlValue,
     Sender,
 };
-use serde::{Deserialize, Serialize};
-use sql_types::SqlType;
 use sqlparser::{
     ast::Statement,
     dialect::{Dialect, PostgreSqlDialect},
@@ -43,43 +41,13 @@ use sqlparser::{
 };
 use std::{iter, sync::Arc};
 
-pub mod catalog_manager;
 mod ddl;
 mod dml;
 mod query;
 mod session;
 
-pub type Projection = (Vec<ColumnDefinition>, Vec<Vec<String>>);
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct ColumnDefinition {
-    name: String,
-    sql_type: SqlType,
-}
-
-impl ColumnDefinition {
-    pub fn new(name: &str, sql_type: SqlType) -> Self {
-        Self {
-            name: name.to_string(),
-            sql_type,
-        }
-    }
-
-    pub fn sql_type(&self) -> SqlType {
-        self.sql_type
-    }
-
-    pub fn has_name(&self, other_name: &str) -> bool {
-        self.name == other_name
-    }
-
-    pub fn name(&self) -> String {
-        self.name.clone()
-    }
-}
-
 pub struct QueryExecutor {
-    storage: Arc<CatalogManager>,
+    storage: Arc<DataManager>,
     sender: Arc<dyn Sender>,
     session: Session,
     processor: QueryProcessor,
@@ -87,7 +55,7 @@ pub struct QueryExecutor {
 }
 
 impl QueryExecutor {
-    pub fn new(storage: Arc<CatalogManager>, sender: Arc<dyn Sender>) -> Self {
+    pub fn new(storage: Arc<DataManager>, sender: Arc<dyn Sender>) -> Self {
         Self {
             storage: storage.clone(),
             sender: sender.clone(),
